+++ packages/block-library/src/navigation/edit/placeholder/index.js
@@ -13,15 +13,14 @@ import useNavigationEntities from '../../use-navigation-entities';
 import PlaceholderPreview from './placeholder-preview';
 import useNavigationMenu from '../../use-navigation-menu';
 import useCreateNavigationMenu from '../use-create-navigation-menu';
-import useConvertClassicMenu from '../../use-convert-classic-menu';
-import ExistingMenusOptions from '../existing-menus-options';
+import NavigationMenuSelector from '../navigation-menu-selector';
 
 export default function NavigationPlaceholder( {
 	clientId,
 	onFinish,
 	canSwitchNavigationMenu,
 	hasResolvedNavigationMenus,
-	canUserCreateNavigation = false,
+	canUserCreateNavigationMenu = false,
 } ) {
 	const createNavigationMenu = useCreateNavigationMenu( clientId );
 
@@ -29,7 +28,7 @@ export default function NavigationPlaceholder( {
 		blocks,
 		navigationMenuTitle = null
 	) => {
-		if ( ! canUserCreateNavigation ) {
+		if ( ! canUserCreateNavigationMenu ) {
 			return;
 		}
 
@@ -40,11 +39,8 @@ export default function NavigationPlaceholder( {
 		onFinish( navigationMenu, blocks );
 	};
 
-	const convertClassicMenu = useConvertClassicMenu( onFinishMenuCreation );
-
 	const {
 		isResolvingPages,
-		menus,
 		isResolvingMenus,
 		hasMenus,
 	} = useNavigationEntities();
@@ -60,7 +56,7 @@ export default function NavigationPlaceholder( {
 	const hasNavigationMenus = !! navigationMenus?.length;
 
 	const showSelectMenus =
-		( canSwitchNavigationMenu || canUserCreateNavigation ) &&
+		( canSwitchNavigationMenu || canUserCreateNavigationMenu ) &&
 		( hasNavigationMenus || hasMenus );
 
 	return (
@@ -94,29 +90,9 @@ export default function NavigationPlaceholder( {
 										popoverProps={ { isAlternate: true } }
 									>
 										{ () => (
-											<ExistingMenusOptions
-												showNavigationMenus={
-													canSwitchNavigationMenu
-												}
-												navigationMenus={
-													navigationMenus
-												}
-												onSelectNavigationMenu={
-													onFinish
-												}
-												classicMenus={ menus }
-												onSelectClassicMenu={ ( {
-													id,
-													name,
-												} ) =>
-													convertClassicMenu(
-														id,
-														name
-													)
-												}
-												showClassicMenus={
-													canUserCreateNavigation
-												}
+											<NavigationMenuSelector
+												clientId={ clientId }
+												onSelect={ onFinish }
 											/>
 										) }
 									</DropdownMenu>
@@ -124,7 +100,7 @@ export default function NavigationPlaceholder( {
 								</>
 							) : undefined }
 
-							{ canUserCreateNavigation && (
+							{ canUserCreateNavigationMenu && (
 								<Button
 									variant="tertiary"
 									onClick={ onCreateEmptyMenu }