+++ packages/data/src/redux-store/metadata/actions.js
@@ -37,10 +37,10 @@ export function finishResolution( selectorName, args ) {
  * started.
  *
  * @param {string}    selectorName Name of selector for which resolver triggered.
- * @param {unknown[]} args         Array of arguments to associate for uniqueness, each item
+ * @param {unknown[][]} args         Array of arguments to associate for uniqueness, each item
  *                                 is associated to a resolution.
  *
- * @return {{ type: 'START_RESOLUTIONS', selectorName: string, args: unknown[] }} Action object.
+ * @return {{ type: 'START_RESOLUTIONS', selectorName: string, args: unknown[][] }} Action object.
  */
 export function startResolutions( selectorName, args ) {
 	return {
@@ -55,10 +55,10 @@ export function startResolutions( selectorName, args ) {
  * completed.
  *
  * @param {string}    selectorName Name of selector for which resolver triggered.
- * @param {unknown[]} args         Array of arguments to associate for uniqueness, each item
+ * @param {unknown[][]} args         Array of arguments to associate for uniqueness, each item
  *                                 is associated to a resolution.
  *
- * @return {{ type: 'FINISH_RESOLUTIONS', selectorName: string, args: unknown[] }} Action object.
+ * @return {{ type: 'FINISH_RESOLUTIONS', selectorName: string, args: unknown[][] }} Action object.
  */
 export function finishResolutions( selectorName, args ) {
 	return {